// src/enums/strava.enums.ts
export enum ActivityMetricAccuracy {
  EXCELLENT = 'excellent',
  HIGH = 'high',
  MEDIUM = 'medium',
  LOW = 'low',
  VERY_LOW = 'very_low',
  NONE = 'none',
}

export enum ActivityMetricSource {
  HEART_RATE = 'heart_rate',
  POWER = 'power',
  KILOJOULES = 'kilojoules',
  CALORIES = 'calories',
  COMPOSITE = 'composite',
  TIME = 'time',
  NONE = 'none',
}